requirements.txt
openai>=1.6.0
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.0
urllib3>=2.0.0
argparse>=1.4.0
flask>=3.1.0 